Updating Github Login Page
