+++ frontend/dev/src/login/Login.css
@@ -104,26 +104,3 @@ i{
 .button .test:hover{
   transform: scale(1.05);
 }
-
-/*Success Text*/
-.success-text {
-  border: none;
-  background-image: grey;
-  color: white;
-  letter-spacing:1px;
-  border-radius: 6px;
-  font-size: 20px;
-  font-weight: 500;
-  text-align: center;
-  padding: 1vmax;
-  margin-bottom: 2vmax;
-
-  cursor: pointer;
-  
-  transition: all 0.3s ease; 
-  animation: 10s rotate linear infinite;
-  background: linear-gradient(45deg, #f04714, #a1134a, #187597, #189276, #c49727,#c52a03);
-  background-size: 600% 100%;
-  animation: gradient 16s linear infinite;
-  animation-direction: alternate;
-}
